Use Sprite size not Tile size when calculating offset of individual sprites.
This doesn't solve all of the problems, but it was one of the bigger issues.
If a sprite was the same size as the tilesize, then it was fine, but otherwise things get weird.

It is important to realize that the "tilesize" is essentially the world's "scale", as in, the pixel to physics "unit" conversion.